requirements.txt
Django==1.8
Pillow==3.1.0
argparse==1.2.1
django-annoying==0.8.5
django-crispy-forms==1.6.0
psycopg2==2.6.1
wsgiref==0.1.2
